+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Added method VApp.AddNewVMWithStorageProfile that adds a VM with custom storage profile.
 * Added command `make static` to run staticcheck on all packages
 * Added `make static` to Travis regular checks
+* Added ability to connect to the vCD using an authorization token
+* Added method `VCDClient.SetToken`
+* Added method `VCDClient.GetAuthResponse`
+* Added script `scripts/get_token.sh`
 
 BUGS FIXED: